backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pymongo==4.7.2
python-dotenv==1.0.1